// src/core/generated/streaming/driver.ts
export interface StreamEvent<TEvent, TState> {
  readonly event: TEvent;
  readonly state: TState;
}